fix: dto priceChange 필드 타입 수정

+++ backend/src/main/java/com/stocktracer/backend/stock/dto/StockPriceDto.java
@@ -19,7 +19,7 @@ public class StockPriceDto {
     private BigDecimal highPrice;
     private BigDecimal lowPrice;
     private Long volume;
-    private float priceChange;
+    private BigDecimal priceChange;
 }
